Move HTTP code out of health.go
Put it with the related HTTP code.
